Don't second-guess version choices
Before, we run '0launch --download-only' to choose some versions and
then recalculated the selections ourself. It was therefore possible that
the versions used didn't match the ones the user had agreed to
(typically because the GUI was using a different version of 0launch).

Now, we use --get-selections to get the selected versions directly.
This requires 0.37, since earlier versions don't include all the
metadata in the selections XML returned.