Don't treat 'src' as a local copy if it coincides with the original
Normally if there is a subdirectory called 'src' we assume it contains
a local copy of the source code and build from that instead. However,
if 'src' is an ancestor or descendant of the upstream source directory
then we're doing a build inside the upstream source code (e.g. a GIT
checkout), so ignore it.