Redesigned setup command
Instead of selecting a set of implementations and saving that for later,
just write a simple configuration file with the name of the interface.

The original plan was to ensure repeatable builds by selecting a version
once and sticking to it. However, it's more common to want to track the
latest version, so that should be the default.