Started adding autocompile command
Given an interface URI, it works out what needs to be compiled next (but
doesn't actually do it yet).

It works as follows:

For every "-src" implementation in every feed used, it pretends that
there is a corresponding binary already compiled for the host platform.
This binary has the same version number and no dependencies. It also has
a fake machine type ("newbuild"), which makes it slightly less desirable
than an existing binary for that version, but more desirable than an
existing binary for an older version.

It then solves as usual, displaying a tree showing all the selected
components. You need to compile and register every selected dependency
which is a source package before compiling the root component itself.