Search legacy packages for /usr/lib/lib*.so broken symlinks and fix
RPM unpacks all packages over the root, so one package can have a symlink to a
file provided by a different package. This is often used to set the default
version of a library in RPM packages:

-devel package:  libSDL.so -> libSDL-1.2.so.0.11.2 (broken link)
runtime package: provides libSDL-1.2.so.0.11.2

Since Zero Install keeps every package in its own directory, this doesn't work.
Instead, search for such broken links, search for a matching target, and create
the fixed link in a temporary directory, added to LD_LIBRARY_PATH.