Added support for <compile:implementation> template
You can put this inside the <compile name='compile'> element to specify a template
for the generated <implementation>. For example, if you are writing a Java program
then you can set the arch to '*-*' to state that the binary works on all platforms.
This also allows adding <command> elements to the binary, and replaces the old
compile:binary-main attribute.